Make mice move in a different manner
